Fix write_build_targets command
The write_build_targets command now first looks for a build_system that
already has the same name. If so, it replaces that build system by the
new build system. If no build system with the same name is present, it
appends the new build system to the list of build systems.

Also, the available commmands that are printed now omit the "_command"
suffix.