Make this server thing somewhat stable
Some problems are still present: The "active target" seems to be
non-deterministic when restoring a session. This is probably due to how Python's
set datastructure works.

Another thing is that the output dir of some targets are not correct. This is
a bug in cmake itself, so I can't fix that. But keep it in mind (note to self).

What else? I've noticed that sometimes the decoding doesn't work and an
exception gets thrown. In this case the server object gets into an unusable
state and you'd have to restart Sublime to get it working again. I should
probably add better exception handling so that the server object can resume
operation as if nothing happened.